fix: solve the dependency injection exception, me route
The issue was that importing providers in the AppModule doesn't
guarantee that they will be accessible for the modules imported in
the AppModule. So I imported the ClientModule.register() directly
in the AuthModule which is where I need it.
After that I made a 'me' GET route that will return the autheniticated
user upon call.